refactor(command-palette): extract groups into command-palette-parts
Was 580 lines doing input row + 3 cmdk groups + footer inline. The 3
groups (Recordings/Actions/Theme) are independent renderings of the
same Row primitive; the footer is purely keyboard hints. Split into:

- command-palette-parts.tsx -- RecordingsGroup, ActionsGroup,
  ThemeGroup, PaletteFooter, plus the Row/Kbd primitives and the
  transcriptSnippet helper (kept in sync with recording-list's local
  copy, see comment).
- command-palette.tsx (189 lines) -- input, cmdk root with controlled
  value + transcribeTargets map for the cmd+enter quick action, and
  composition of the parts.

The transcribeTargets logic stays in the parent because it depends on
activeValue (cmdk's controlled state) and the keydown handler that
reads it. Pulling it into a part would require either lifting cmdk
state up too or threading a ref through the boundary -- not worth it.

No behavior change.
